Fix package.json for correct publishing and Node.js compat
- Point "main" at ESM source instead of nonexistent UMD bundle path
- Add "exports" field with types condition
- Add "files" field and remove .npmignore (redundant)
- Replace deprecated "prepublish" with "prepublishOnly"
- Fix webpack globalObject for Node.js UMD compatibility
- Inline AudioWorklet code in speakers.js to remove the ?raw bundler
  dependency, which eliminates the need for a separate Node.js entry
  point

The UMD dist bundles are for <script> tags and CDNs only. All Node.js
and bundler consumers get the ESM source via "exports" / "main", which
Node.js 22 handles natively for both import and require().
