--- .npmignore
@@ -1,2 +1,0 @@
-/local-roms
-/tmp
