--- src/browser/audio-worklet-processor.js
@@ -1,75 +1,0 @@
-// AudioWorklet processor for JSNES audio output.
-// Receives stereo samples from the main thread via MessagePort and
-// buffers them in a circular Float32Array for playback in process().
-
-class NESAudioProcessor extends AudioWorkletProcessor {
-  constructor() {
-    super();
-    // Circular buffer sized to hold ~170ms of audio at 48kHz (8192 samples).
-    // Large enough to absorb timing jitter between the main thread and the
-    // audio thread, but small enough to keep latency low.
-    this.capacity = 8192;
-    this.bufferL = new Float32Array(this.capacity);
-    this.bufferR = new Float32Array(this.capacity);
-    this.readPos = 0;
-    this.writePos = 0;
-    this.count = 0;
-
-    this.port.onmessage = (e) => {
-      if (e.data.type === "samples") {
-        const left = e.data.left;
-        const right = e.data.right;
-        const len = left.length;
-
-        // If adding these samples would overflow, drop oldest to make room
-        if (this.count + len > this.capacity) {
-          const drop = this.count + len - this.capacity;
-          this.readPos = (this.readPos + drop) % this.capacity;
-          this.count -= drop;
-        }
-
-        for (let i = 0; i < len; i++) {
-          this.bufferL[this.writePos] = left[i];
-          this.bufferR[this.writePos] = right[i];
-          this.writePos = (this.writePos + 1) % this.capacity;
-        }
-        this.count += len;
-      }
-    };
-  }
-
-  process(inputs, outputs) {
-    const output = outputs[0];
-    if (!output || output.length < 2) return true;
-
-    const outL = output[0];
-    const outR = output[1];
-    const size = outL.length;
-
-    if (this.count < size) {
-      // Buffer underrun — output what we have, pad rest with silence
-      for (let i = 0; i < this.count; i++) {
-        outL[i] = this.bufferL[this.readPos];
-        outR[i] = this.bufferR[this.readPos];
-        this.readPos = (this.readPos + 1) % this.capacity;
-      }
-      for (let i = this.count; i < size; i++) {
-        outL[i] = 0;
-        outR[i] = 0;
-      }
-      this.count = 0;
-      this.port.postMessage({ type: "underrun" });
-    } else {
-      for (let i = 0; i < size; i++) {
-        outL[i] = this.bufferL[this.readPos];
-        outR[i] = this.bufferR[this.readPos];
-        this.readPos = (this.readPos + 1) % this.capacity;
-      }
-      this.count -= size;
-    }
-
-    return true;
-  }
-}
-
-registerProcessor("nes-audio-processor", NESAudioProcessor);
